requirements.txt
cmake
dlib==19.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.6
Flask-Mail==0.9.1
Flask-SocketIO==3.0.2
gunicorn==19.9.0
opencv-python==3.4.3.18
pandas==0.23.4
redis==2.10.6
rq==0.12.0
scipy==1.1.0
slidingwindow==0.0.13
tensorflow==1.11.0
google-cloud-pubsub==0.38.0
requests
xgboost
sklearn
tensorflow-gpu
argparse
dill
fire
matplotlib
psutil
requests
scikit-image
tqdm
git+https://github.com/ppwwyyxx/tensorpack.git
















